+++ rss-react-app/src/services/api.ts
@@ -1,4 +1,4 @@
-import { Character } from '../components/interfaces';
+import { Character } from '../types/interfaces';
 
 const homeworldCache: { [url: string]: string } = {};
