+++ rss-react-app/src/styles/DetailedCard.css
