+++ rss-react-app/src/styles/SearchResults.css
