+++ rss-react-app/src/types/interfaces.tsx
